+++ src/sdc11073/consumer/consumerimpl.py
@@ -2,13 +2,15 @@
 from __future__ import annotations
 
 import copy
+import functools
 import logging
 import ssl
+import time
 import traceback
 import uuid
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any
-from urllib.parse import urlparse, urlsplit
+from urllib.parse import urlparse
 
 from lxml import etree as etree_
 
@@ -31,18 +33,19 @@ from .request_handler_deferred import EmptyResponse
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
-    from sdc11073.dispatch.request import RequestData
-    from sdc11073.xml_types.mex_types import HostedServiceType
-    from sdc11073.pysoap.soapclient import SoapClientProtocol
-    from sdc11073.pysoap.msgreader import MessageReader, ReceivedMessage
-    from sdc11073.pysoap.msgfactory import MessageFactory
-    from sdc11073.definitions_base import AbstractDataModel, BaseDefinitions
-    from sdc11073.mdib.consumermdib import ConsumerMdib
+
     from sdc11073.consumer.serviceclients.serviceclientbase import HostedServiceClient
     from sdc11073.consumer.subscription import ConsumerSubscriptionManagerProtocol
+    from sdc11073.definitions_base import AbstractDataModel, BaseDefinitions
+    from sdc11073.dispatch.request import RequestData
+    from sdc11073.mdib.consumermdib import ConsumerMdib
+    from sdc11073.pysoap.msgfactory import MessageFactory
+    from sdc11073.pysoap.msgreader import MessageReader, ReceivedMessage
+    from sdc11073.pysoap.soapclient import SoapClientProtocol
     from sdc11073.wsdiscovery.service import Service
-    from .components import SdcConsumerComponents
+    from sdc11073.xml_types.mex_types import HostedServiceType
 
+    from .components import SdcConsumerComponents
     from .subscription import ConsumerSubscription
 
 
@@ -145,7 +148,7 @@ class _NotificationsSplitter:
             actions.PeriodicContextReport: 'periodic_context_report',
             actions.DescriptionModificationReport: 'description_modification_report',
             actions.OperationInvokedReport: 'operation_invoked_report',
-            actions.SystemErrorReport: 'system_error_report'
+            actions.SystemErrorReport: 'system_error_report',
         }
 
 
@@ -157,7 +160,9 @@ class SdcConsumer:
     What if not???? => raise exception in _discover_hosted_services.
     """
 
-    is_connected = properties.ObservableProperty(False)  # a boolean
+    subscription_status = properties.ObservableProperty({})
+    # indicates whether all subscriptions have been subscribed to and are renewed successfully
+    is_connected = properties.ObservableProperty(False)
 
     # observable properties for all notifications
     # all incoming Notifications can be observed in state_event_report ( as soap envelope)
@@ -192,10 +197,13 @@ class SdcConsumer:
                  default_components: SdcConsumerComponents | None = None,
                  specific_components: SdcConsumerComponents | None = None,
                  request_chunk_size: int = 0,
-                 socket_timeout: int = 5):
+                 socket_timeout: int = 5,
+                 force_ssl_connect: bool = False,
+                 ):
         """Construct a SdcConsumer.
 
-        :param device_location: the XAddr location for meta data, e.g. http://10.52.219.67:62616/72c08f50-74cc-11e0-8092-027599143341
+        :param device_location: the XAddr location for meta data,
+                                e.g. http://10.52.219.67:62616/72c08f50-74cc-11e0-8092-027599143341
         :param sdc_definitions: a class derived from BaseDefinitions
         :param epr: the path of this client in http server
         :param ssl_context_container: used for ssl connection to device and for own HTTP Server (notifications receiver)
@@ -204,9 +212,24 @@ class SdcConsumer:
         :param specific_components: a SdcConsumerComponents instance or None
         :param request_chunk_size: if value > 0, message is split into chunks of this size
         :param socket_timeout: timeout for connections to provider
+        :param force_ssl_connect: True: only accept ssl connections (requires a ssl_context_container)
+                                  False: if ssl_context_container is provided, consumer first tries an
+                                         encrypted connection, and if this raises an SSLError,
+                                         it tries an unencrypted connection
         """
         if not device_location.startswith('http'):
             raise ValueError('Invalid device_location, it must be match http(s)://<netloc> syntax')
+        self.is_ssl_connection: bool | None
+        if force_ssl_connect:
+            if ssl_context_container is None:
+                raise ValueError(
+                    'Invalid combination of ssl_connect (True) and ssl_context_container (None) parameters')
+            self.is_ssl_connection = True
+        elif ssl_context_container is None:
+            self.is_ssl_connection = False
+        else:
+            self.is_ssl_connection = None  # options allow both, needs to be decided when connecting
+
         self._device_location = device_location
         self.sdc_definitions = sdc_definitions
         if default_components is None:
@@ -214,8 +237,8 @@ class SdcConsumer:
         self._components = copy.deepcopy(default_components)
         if specific_components is not None:
             self._components.merge(specific_components)
-        splitted = urlsplit(self._device_location)
-        self._device_uses_https = splitted.scheme.lower() == 'https'
+
+        self.subscription_status: dict[str, bool] = {}
 
         # available after start_all
         self._network_adapter: network.NetworkAdapter | None = None
@@ -314,6 +337,8 @@ class SdcConsumer:
 
         Replace servers ip address with own ip address (server might have 0.0.0.0).
         """
+        if self._http_server is None:
+            return ''
         p = urlparse(self._http_server.base_url)
         tmp = f'{p.scheme}://{self._network_adapter.ip}:{p.port}{p.path}'
         sep = '' if tmp.endswith('/') else '/'
@@ -332,6 +357,15 @@ class SdcConsumer:
         :return: a subscription object.
         """
         subscription = self._subscription_mgr.mk_subscription(dpws_hosted, filter_type)
+
+        def update_subscription_status(subscription_filter: str, status: bool):
+            subscription_status = dict(self.subscription_status)
+            subscription_status[subscription_filter] = status
+            self.subscription_status = subscription_status  # trigger observable if status has changed
+
+        properties.strongbind(subscription, is_subscribed=functools.partial(update_subscription_status,
+                                                                            filter_type.text))
+
         # direct subscribed notifications to this subscription
         for action in actions:
             self._services_dispatcher.register_post_handler(action, subscription.on_notification)
@@ -438,9 +472,10 @@ class SdcConsumer:
                which is the minimal requirement for a sdc provider.
         :return: None
         """
-        if self.host_description is None:
-            self._logger.debug('reading meta data from {}', self._device_location)  # noqa: PLE1205
-            self.host_description = self._get_metadata()
+        self._logger.debug('connecting to %s', self._device_location)
+        self._connect()
+        self._logger.debug('reading meta data from %s', self._device_location)
+        self.host_description = self._get_metadata()
 
         # now query also metadata of hosted services
         self._mk_hosted_services(self.host_description)
@@ -457,11 +492,6 @@ class SdcConsumer:
             raise RuntimeError(f'GetService not detected! found services = {list(self._service_clients.keys())}')
 
         self._start_event_sink(shared_http_server)
-        periodic_actions = {self.sdc_definitions.Actions.PeriodicMetricReport,
-                            self.sdc_definitions.Actions.PeriodicAlertReport,
-                            self.sdc_definitions.Actions.PeriodicComponentReport,
-                            self.sdc_definitions.Actions.PeriodicContextReport,
-                            self.sdc_definitions.Actions.PeriodicOperationalStateReport}
 
         # start subscription manager
         subscription_manager_class = self._components.subscription_manager_class
@@ -513,18 +543,17 @@ class SdcConsumer:
                         self._logger.error('start_all: could not subscribe: error = {}, actions= {}',  # noqa: PLE1205
                                            traceback.format_exc(), subscribe_actions)
 
+        def _update_is_connected(subscription_status: dict[str, bool]):
+            self.is_connected = all(subscription_status.values()) and any(subscription_status)
+
+        properties.strongbind(self, subscription_status=_update_is_connected)
+        _update_is_connected(self.subscription_status)
+
         # register callback for end of subscription
         self._services_dispatcher.register_post_handler(
             DispatchKey(EventingActions.SubscriptionEnd,
                         self.sdc_definitions.data_model.ns_helper.WSE.tag('SubscriptionEnd')),
             self._on_subscription_end)
-
-        # connect self.is_connected observable to all_subscriptions_okay observable in subscriptions manager
-        def set_is_connected(is_ok: bool):
-            self.is_connected = is_ok
-
-        properties.strongbind(self._subscription_mgr, all_subscriptions_okay=set_is_connected)
-        self.is_connected = self._subscription_mgr.all_subscriptions_okay
 
     def stop_all(self, unsubscribe: bool = True):
         """Stop all threads, optionally unsubscribe."""
@@ -544,24 +573,38 @@ class SdcConsumer:
         del self._compression_methods[:]
         self._compression_methods.extend(compression_methods)
 
-    def _get_metadata(self) -> mex_types.Metadata:
-        _url = urlparse(self._device_location)
-        wsc = self.get_soap_client(self._device_location)
-
-        if self._ssl_context_container is not None and _url.scheme == 'https':
-            if wsc.is_closed():
-                wsc.connect()
-            sock = wsc.sock
+    def _connect(self):
+        soap_client = self.get_soap_client(self._device_location)
+        if self.is_ssl_connection is not None:
+            # decision was already made in constructor
+            soap_client.connect()
+        else:
+            try:
+                soap_client.connect()
+                self.is_ssl_connection = True
+            except ssl.SSLError:
+                # could not connect with ssl, try without it
+                soap_client.close()
+                self._forget_soap_client(soap_client)
+                self.is_ssl_connection = False
+                soap_client = self.get_soap_client(self._device_location)
+                # if this also fails, something else is wrong and error needs handling on application level.
+                soap_client.connect()
+        if self.is_ssl_connection:
+            sock = soap_client.sock
             self.peer_certificate = sock.getpeercert(binary_form=False)
             self.binary_peer_certificate = sock.getpeercert(binary_form=True)  # in case the application needs it...
-
             self._logger.info('Peer Certificate: {}', self.peer_certificate)  # noqa: PLE1205
+
+    def _get_metadata(self) -> mex_types.Metadata:
+        _url = urlparse(self._device_location)
+        soap_client = self.get_soap_client(self._device_location)
         nsh = self.sdc_definitions.data_model.ns_helper
         inf = HeaderInformationBlock(action=f'{nsh.WXF.namespace}/Get',
                                      addr_to=self._device_location)
         message = self.msg_factory.mk_soap_message_etree_payload(inf, payload_element=None)
 
-        received_message_data = wsc.post_message_to(_url.path, message, msg='getMetadata')
+        received_message_data = soap_client.post_message_to(_url.path, message, msg='getMetadata')
         return mex_types.Metadata.from_node(received_message_data.p_msg.body_node)
 
     def send_probe(self) -> ProbeMatchesType:
@@ -579,20 +622,25 @@ class SdcConsumer:
     def get_soap_client(self, address: str) -> SoapClientProtocol:
         """Return the soap client for address.
 
-        Method creates a new soap client if needed.
+        Method creates a new soap client if needed and considers self.is_ssl_connection value.
         """
         _url = urlparse(address)
-        key = (_url.scheme, _url.netloc)
+        use_ssl = self.is_ssl_connection is not False  # if is_ssl_connection is still None, default to use_ssl = True
+        key = (use_ssl, _url.netloc)
         soap_client = self._soap_clients.get(key)
         if soap_client is None:
-            soap_client = self._mk_soap_client(_url.scheme, _url.netloc)
+            soap_client = self._mk_soap_client(use_ssl, _url.netloc)
             self._soap_clients[key] = soap_client
         return soap_client
 
-    def _mk_soap_client(self, scheme: str,  # noqa: PLR0913
-                        netloc: str) -> SoapClientProtocol:
-        _ssl_context = \
-            self._ssl_context_container.client_context if scheme == "https" and self._ssl_context_container else None
+    def _forget_soap_client(self, soap_client: SoapClientProtocol):
+        for key, value in self._soap_clients.items():
+            if value is soap_client:
+                del self._soap_clients[key]
+                return
+
+    def _mk_soap_client(self, use_ssl: bool, netloc: str) -> SoapClientProtocol:
+        _ssl_context = self._ssl_context_container.client_context if use_ssl else None
         cls = self._components.soap_client_class
         return cls(netloc,
                    self._socket_timeout,
@@ -633,7 +681,7 @@ class SdcConsumer:
     def _start_event_sink(self, shared_http_server: Any):
         if shared_http_server is None:
             self._is_internal_http_server = True
-            ssl_context_container = self._ssl_context_container if self._device_uses_https else None
+            ssl_context_container = self._ssl_context_container if self.is_ssl_connection else None
             logger = loghelper.get_logger_adapter('sdc.client.notif_dispatch', self.log_prefix)
             self._http_server = HttpServerThreadBase(
                 str(self._network_adapter.ip),
@@ -643,6 +691,9 @@ class SdcConsumer:
             )
             self._http_server.start()
             self._http_server.started_evt.wait(timeout=5)
+            # it sometimes still happens that http server is not completely started without waiting.
+            # TODO: find better solution, see issue #320
+            time.sleep(1)
             self._logger.info('serving EventSink on {}', self._http_server.base_url)  # noqa: PLE1205
         else:
             self._http_server = shared_http_server